made text extraction run from model dependency injection rather than hardcoded dependency, wrote an OllamaLLM and LLM registry
