+++ server/src/infra/llm/registry.llm.js
@@ -1,0 +1,10 @@
+import { GroqLLM } from "./groq.llm.js";
+import { OllamaLLM } from "./ollama.llm.js";
+
+const gq = new GroqLLM();
+const ol = new OllamaLLM();
+
+export const ModelRegistry = {
+    'GROQ':   gq,
+    'Ollama': ol
+};
