+++ server/src/services/resume.service.js
@@ -1,41 +1,30 @@
 class ResumeService {
 
     // takes no dependencies at instantiation
-    constructor(
-        model,
-        ParsingService,
-        EmbeddingService,
-    ) {
+    // instead uses dependency injection at runtime
+    constructor() {}
 
-        if (null in [model, ParsingService, EmbeddingService])
-            throw new Error('One or more inputs to ResumeService were null');
 
-        this.model    = model;
-        this.parser   = ParsingService;
-        this.embedder = EmbeddingService;
-        this.chunks   = ChunksModel;
-    }
-
-    async getNewResumes(ResumeSource) {
+    async getNewResumes(ResumeProvider) {
         
         // get the ids of resumes in the collection
         const resumes = await this.model.find();
         const oldIds = resumes.map(resume => resume.sourceId.value);
 
         // pass the oldIds as a filter to the ResumeSource
-        const newResumes = await ResumeSource.getResumes(oldIds);
-        return newResumes;        
+        const newResumes = await ResumeProvider.getResumes(oldIds);
+        return newResumes;
     }
 
-    async saveToDatabase(resumes) {
+    async saveToDatabase(resumes, database) {
         if (Array.isArray(resumes))
-            await this.model.bulkWrite(resumes);
+            await database.bulkWrite(resumes);
         else
-            await this.model.insert(resumes);
+            await database.insert(resumes);
     }
 
-    async chunkResumes(resumes) {
-        
+    async chunkResumes(ParsingService) {
+        const TextChunker = ParsingService.
     }
 
 };
